scout.py
import feedparser
import requests
from bs4 import BeautifulSoup
from newspaper import Article
import fastf1
import pandas as pd
import json
import logging
from datetime import datetime

# Setup logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger("Scout")

class Scout:
    def __init__(self):
        self.rss_feeds = [
            "https://www.motorsport.com/rss/f1/news/",
            "https://www.autosport.com/rss/feed/f1",
            "https://www.motorsport.com/rss/motogp/news/",
            "https://www.autosport.com/rss/feed/motogp"
        ]
        # Enable FastF1 cache - assuming a default location or user config
        # fastf1.Cache.enable_cache('path/to/cache') # Uncomment and set path if needed

    def fetch_news(self):
        """
        Fetches latest news from RSS feeds.
        Filters for 'Breaking', 'Results', 'Driver Transfers' logic to be improved.
        """
        logger.info("Fetching news from RSS feeds...")
        news_items = []
        
        for url in self.rss_feeds:
            feed = feedparser.parse(url)
            for entry in feed.entries[:5]: # Check top 5 from each
                # Basic filtering logic (can be expanded)
                title = entry.title
                link = entry.link
                image_url = self._extract_image(entry)
                
                # Check for keywords
                title_lower = title.lower()
                item_type = "NEWS" # Default
                
                if "result" in title_lower or "qualifying" in title_lower or "practice" in title_lower or "winner" in title_lower:
                    item_type = "RESULT"
                elif "transfer" in title_lower or "sign" in title_lower or "contract" in title_lower:
                     item_type = "OFFICIAL"
                elif "rumour" in title_lower or "report" in title_lower or "suggests" in title_lower or "could" in title_lower:
                     item_type = "RUMOUR"
                elif "breaking" in title_lower:
                     item_type = "BREAKING"
                elif "analysis" in title_lower or "tech" in title_lower:
                     item_type = "ANALYSIS"
                
                # Clean summary
                summary = ""
                # Clean summary
                summary = ""
                try:
                    # Attempt full text extraction
                    article = Article(link)
                    article.download()
                    article.parse()
                    
                    # Use the first 500-600 characters but try to end on a full sentence
                    full_text = article.text.strip()
                    if len(full_text) > 600:
                        summary = full_text[:600].rsplit('.', 1)[0] + "."
                    else:
                        summary = full_text
                except Exception as e:
                    logger.warning(f"Scraping failed for {link}: {e}")

                if not summary and 'summary' in entry:
                    # Fallback to RSS summary
                    soup = BeautifulSoup(entry.summary, "html.parser")
                    summary = soup.get_text()[:400] + "..."
                
                # Create item
                item = {
                    "id": entry.id if 'id' in entry else link,
                    "headline_en": title,
                    "headline_ta": self.translate_headline(title),
                    "summary": summary,
                    "image_url": image_url,
                    "link": link,
                    "type": item_type,
                    "source": "RSS"
                }
                news_items.append(item)
                
        return news_items

    def _extract_image(self, entry):
        """
        Attempt to find an image in the RSS entry.
        """
        if 'media_content' in entry:
            return entry.media_content[0]['url']
        if 'links' in entry:
            for l in entry.links:
                 if 'image' in l.type:
                     return l.href
        return None

    def fetch_telemetry(self, year=None, gp=None, session='R'):
        """
        Fetches race results using FastF1.
        Defaults to latest race if not specified.
        """
        logger.info("Fetching telemetry/results...")
        if not year or not gp:
            # Logic to find latest race could happen here
            # For now, let's just return a placeholder or specific race for testing
            # Getting the schedule for current year
            try:
                current_year = datetime.now().year
                schedule = fastf1.get_event_schedule(current_year)
                # Logic to find last completed race
                # This is complex, for MVP we might manual trigger or just check recent
                pass
            except Exception as e:
                logger.error(f"Error fetching schedule: {e}")
                return []

        # Example: Return dummy result data structure for now
        # In a real scenario, we would `fastf1.get_session(year, gp, session).load()`
        return []

    def translate_headline(self, text):
        """
        Placeholder for Tamil translation.
        """
        # TODO: Integrate with a translation API (Google Translate, DeepL, etc.)
        return f"[TA] {text}"

if __name__ == "__main__":
    scout = Scout()
    items = scout.fetch_news()
    print(json.dumps(items, indent=2))
