requirements.txt
# --- Agent Alpha: The Scout (News & Data) ---
feedparser>=6.0.10
beautifulsoup4>=4.12.0
fastf1>=3.1.0
pandas>=2.1.0
newspaper3k>=0.2.8
lxml[html_clean]>=5.2.0
pandas>=2.1.0

# --- Agent Beta: The Studio (Image & Video) ---
Pillow>=10.0.0
moviepy>=1.0.3
numpy>=1.26.0
edge-tts>=6.1.9

# --- Agent Gamma: The Publisher (Social Media) ---
instagrapi>=1.17.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# --- System & Utilities ---
python-dotenv>=1.0.0
schedule>=1.2.0
tqdm>=4.66.0
pytest>=7.0.0
streamlit>=1.31.0
requests-toolbelt>=1.0.0
zstandard>=0.19.0
